resolved conflicts with Results class

+++ game/Results.java
@@ -6,7 +6,7 @@ public class Results {
     public int[] state;
     public boolean terminated;
 
-    public Results(final int rew, final int[] sta, final boolean term) {
+    public Results(final double rew, final int[] sta, final boolean term) {
         reward = rew; state = sta; terminated = term;
     }
 }
